Updates to NKI docs and whats new page

+++ nki/examples/matrix_multiplication/matrix_multiplication_nki_kernels.py
@@ -59,7 +59,7 @@ def nki_matmul_basic_(lhsT, rhs):
   # Create a tensor in SBUF and copy the result from PSUM back to SBUF, 
   # and cast to expected output data-type
   result_sbuf = nl.ndarray(result_psum.shape, dtype=result.dtype, buffer=nl.sbuf)
-  nisa.tensor_copy(dst=result_sbuf, src=result_psum, dtype=result.dtype)
+  nisa.tensor_copy(dst=result_sbuf, src=result_psum)
 
   # The result of [64,128] x [128,512] matrix multiplication has a shape of [64, 512].
   # This dictates which indices to use to address the result tile.
@@ -130,7 +130,7 @@ def nki_matmul_tiled_(lhsT, rhs):
 
       # Copy the result from PSUM back to SBUF, and cast to expected output data-type
       res_sb = nl.ndarray(res_psum.shape, dtype=result.dtype, buffer=nl.sbuf)
-      nisa.tensor_copy(dst=res_sb, src=res_psum, dtype=result.dtype)
+      nisa.tensor_copy(dst=res_sb, src=res_psum)
 
       # Copy the result from SBUF to HBM.
       nisa.dma_copy(dst=result[m * TILE_M:(m + 1) * TILE_M,
@@ -215,7 +215,7 @@ def nki_matmul_hoist_load_(lhsT, rhs):
 
       # Copy the result from PSUM back to SBUF, and cast to expected output data-type
       res_sb = nl.ndarray(shape=(TILE_M, TILE_N), dtype=nl.float32, buffer=nl.sbuf)
-      nisa.tensor_copy(dst=res_sb, src=res_psum, dtype=result.dtype)
+      nisa.tensor_copy(dst=res_sb, src=res_psum)
 
       # Copy the result from SBUF to HBM.
       nisa.dma_copy(dst=result[m * TILE_M:(m + 1) * TILE_M,
